Port to elm 0.18.0
