+++ elm-package.json
@@ -6,9 +6,11 @@
     "source-directories": [
         "src"
     ],
-    "exposed-modules": ["Complex"],
+    "exposed-modules": [
+        "Complex"
+    ],
     "dependencies": {
-        "elm-lang/core": "4.0.0 <= v < 5.0.0"
+        "elm-lang/core": "5.1.1 <= v < 6.0.0"
     },
-    "elm-version": "0.17.0 <= v < 0.18.0"
+    "elm-version": "0.18.0 <= v < 0.19.0"
 }
