--- tests/elm-package.json
@@ -1,15 +1,0 @@
-{
-    "version": "1.0.0",
-    "summary": "helpful summary of your project, less than 80 characters",
-    "repository": "https://github.com/user/project.git",
-    "license": "BSD3",
-    "source-directories": [
-        ".","../src"
-    ],
-    "exposed-modules": [],
-    "dependencies": {
-        "elm-community/elm-test": "1.1.0 <= v < 2.0.0",
-        "elm-lang/core": "4.0.0 <= v < 5.0.0"
-    },
-    "elm-version": "0.17.0 <= v < 0.18.0"
-}
